Add files via upload
4 tests failed

Test Case
Target received damage, but it shouldn't.
Test Case
Damage dealt to the enemy target after using an ability is incorrect.
Test Case
Could not instantiate the class Mage because the constructor with 5 parameters threw an exception and could not be initialized. Make sure to check the constructor implementation.
Test Case
Expected output: Test has no skills to use!